+++ src/main/java/com/finalfantasy/football/players/repositories/KickerRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface KickerRepository extends JpaRepository<Kicker, Integer> {
+public interface KickerRepository extends JpaRepository<Kicker, Long> {
 }
