+++ src/main/java/com/finalfantasy/football/players/repositories/QuarterbackRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface QuarterbackRepository extends JpaRepository<Quarterback, Integer> {
+public interface QuarterbackRepository extends JpaRepository<Quarterback, Long> {
 }
